Revert "feat(vimeo): Add Vimeo API / loader"
