+++ src/config.ts
@@ -81,8 +81,6 @@ const {
   TINEYE_API_PASSWORD,
   TINEYE_API_USERNAME,
   UNLEASH_ADMIN_TOKEN,
-  VIMEO_API_BASE,
-  VIMEO_API_TOKEN,
   VORTEX_API_BASE,
   VORTEX_APP_ID,
   VORTEX_TOKEN,
@@ -219,8 +217,6 @@ export default {
   TINEYE_API_PASSWORD,
   TINEYE_API_USERNAME,
   UNLEASH_ADMIN_TOKEN,
-  VIMEO_API_BASE,
-  VIMEO_API_TOKEN,
   VORTEX_API_BASE,
   VORTEX_APP_ID,
   VORTEX_TOKEN,
