+++ src/lib/loaders/api/index.ts
@@ -12,7 +12,6 @@ import { vortex } from "lib/apis/vortex"
 import { greenhouse } from "lib/apis/greenhouse"
 import { ipbase } from "lib/apis/ipbase"
 import { unleash } from "lib/apis/unleash"
-import { vimeo } from "lib/apis/vimeo"
 
 import { apiLoaderWithAuthenticationFactory } from "lib/loaders/api/loader_with_authentication_factory"
 import { apiLoaderWithoutAuthenticationFactory } from "lib/loaders/api/loader_without_authentication_factory"
@@ -174,20 +173,6 @@ export default (opts) => ({
   ),
 
   /**
-   * The Vimeo loaders produced by this factory _will_ cache all responses to memcache.
-   *
-   * Do **not** use it for authenticated requests!
-   */
-  vimeoLoaderWithoutAuthenticationFactory: apiLoaderWithoutAuthenticationFactory(
-    vimeo,
-    "vimeo",
-    {
-      requestIDs: opts.requestIDs,
-      userAgent: opts.userAgent,
-    }
-  ),
-
-  /**
    * The Gravity loaders produced by this factory _will_ cache responses for the duration of query execution but do
    * **not** cache to memcache.
    *
@@ -226,21 +211,6 @@ export default (opts) => ({
   unleashLoaderWithAuthenticationFactory: apiLoaderWithAuthenticationFactory(
     unleash,
     "unleash",
-    {
-      requestIDs: opts.requestIDs,
-      userAgent: opts.userAgent,
-    }
-  ),
-
-  /**
-   * The Vimeo loaders produced by this factory _will_ cache responses for the duration of query execution but do
-   * **not** cache to memcache.
-   *
-   * Use this for authenticated requests.
-   */
-  vimeoLoaderWithAuthenticationFactory: apiLoaderWithAuthenticationFactory(
-    vimeo,
-    "vimeo",
     {
       requestIDs: opts.requestIDs,
       userAgent: opts.userAgent,
