+++ src/lib/loaders/loaders_without_authentication/index.ts
@@ -7,7 +7,6 @@ import gravityLoaders from "./gravity"
 import positronLoaders from "./positron"
 import greenhouseLoaders from "./greenhouse"
 import ipbaseLoaders from "./ipbase"
-import { vimeoLoaders } from "./vimeo"
 import vortexLoaders from "./vortex"
 
 export const createLoadersWithoutAuthentication = (opts) => ({
@@ -20,7 +19,6 @@ export const createLoadersWithoutAuthentication = (opts) => ({
   ...geodataLoaders(opts),
   ...greenhouseLoaders(opts),
   ...ipbaseLoaders(opts),
-  ...vimeoLoaders(opts),
   ...vortexLoaders(opts),
 })
